requirements-prod.txt
algoliasearch==2.4.0
aniso8601==8.0.0
apipkg==1.5
appdirs==1.4.4
argcomplete==1.12.0
arrow==0.15.8
asgiref==3.2.10
asn1crypto==1.4.0
atomicwrites==1.4.0
attrs==19.3.0
auger-python==0.1.35
better-exceptions==0.2.2
blessings==1.7
blinker==1.4
boto3==1.16.30
botocore==1.19.30
cachetools==4.1.1
certifi==2020.6.20
cfn-flip==1.2.3
chardet==3.0.4
Click==7.1.2
colorclass==2.2.0
coreapi==2.3.3
coreschema==0.0.4
curtsies==0.3.4
daiquiri==2.1.1
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1
django-code-generator==0.2.2
django-cors-headers==3.4.0
django-filter==2.3.0
django-health-check==3.12.1
django-redis==4.12.1
django-s3-storage==0.13.3
djangorestframework==3.11.1
docopt==0.6.2
docutils==0.16
drf-yasg==1.17.1
ecdsa==0.14.1
entrypoints==0.3
execnet==1.7.1
Faker==0.9.1
fire==0.3.1
Flask==1.1.2
funcsigs==1.0.2
future==0.18.2
geographiclib==1.50
geopy==2.0.0
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
interrogate==1.2.0
itsdangerous==1.1.0
itypes==1.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
mixpanel==4.6.0
more-itertools==8.4.0
networkx==2.4
packaging==20.4
parso==0.7.1
pathspec==0.8.0
placebo==0.9.0
pluggy==0.13.1
prompt-toolkit==3.0.6
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.0.2
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
python-json-logger==0.1.11
python-slugify==3.0.2
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
rope==0.17.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
sentry-sdk==0.19.3
simplejson==3.17.2
six==1.15.0
snowballstemmer==2.0.0
sqlparse==0.3.1
tabulate==0.8.7
terminaltables==3.1.0
text-unidecode==1.2
toml==0.10.1
tqdm==4.48.2
typed-ast==1.4.1
Unidecode==1.1.1
uritemplate==3.0.1
wcwidth==0.2.5
zappa==0.52.0
zipp==3.1.0
trafaret==2.0.2
python-jose==3.2.0
django-server-timing==0.0.3
django-prometheus==2.1.0
